requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
python-docx
google-genai
pydantic
python-dotenv